+++ src/http/oauth1.ts
@@ -133,7 +133,7 @@ export function createOAuth1HttpClient(
       secret: tokenResult.data.tokenSecret,
     };
 
-    setTokenData(tokenResultData);
+    await setTokenData(tokenResultData);
     return tokenResultData;
   }
 
@@ -168,7 +168,7 @@ export function createOAuth1HttpClient(
       secret: tokenResult.data.tokenSecret,
     };
 
-    setTokenData(tokenResultData);
+    await setTokenData(tokenResultData);
     return tokenResultData;
   }