+++ src/services/data/types.ts
@@ -5,6 +5,7 @@ import type {
   OptionsBase,
   OptionsWithRql,
   PagedResult,
+  PagedResultWithPager,
 } from '../types';
 
 export enum JSONSchemaType {
@@ -481,6 +482,7 @@ export interface DataDocumentsService {
    * @param schemaId the schema Id
    * @param documentId the document Id
    * @returns boolean success
+   * @throws {Error} If the document is in a locked state after the specified tries
    */
   assertNonLockedState(
     schemaId: ObjectId,
@@ -533,12 +535,12 @@ export interface DataDocumentsService {
    * `VIEW_DOCUMENTS` | `global` | See any document
    * @param schemaId The id of the targeted schema.
    * @param rql Add filters to the requested list.
-   * @returns {Document} document
+   * @returns PagedResultWithPager<Document>
    */
   find<CustomData = null>(
     schemaId: ObjectId,
     options?: OptionsWithRql
-  ): Promise<PagedResult<Document<CustomData>>>;
+  ): Promise<PagedResultWithPager<Document<CustomData>>>;
   /**
    * Request a list of all documents
    *
@@ -573,7 +575,38 @@ export interface DataDocumentsService {
     schemaId: ObjectId,
     options?: OptionsWithRql
   ): Promise<Document<CustomData>[]>;
-
+  /**
+   * Request a list of all documents and return a generator
+   *
+   * ReadMode on schema is set to 'default' (or the property is not set at all on the schema):
+   *
+   * Permission | Scope | Effect
+   * - | - | -
+   * none | | See your own documents
+   * none | `staff enlistment` | See the documents belonging to the group (and your own documents)
+   * `VIEW_DOCUMENTS` | `global` | See any document
+   *
+   * ReadMode on schema is set to 'allUsers':
+   *
+   * Permission | Scope | Effect
+   * - | - | -
+   * none | | See any document
+   *
+   * ReadMode on schema is set to 'enlistedInLinkedGroups':
+   *
+   * Permission | Scope | Effect
+   * - | - | -
+   * none | `patient enlistment` | See the documents belonging to the group
+   * none | `staff enlistment` | See the documents belonging to the group
+   * `VIEW_DOCUMENTS` | `global` | See any document
+   * @param schemaId The id of the targeted schema.
+   * @param rql Add filters to the requested list.
+   * @returns Document[]
+   */
+  findAllIterator<CustomData = null>(
+    schemaId: ObjectId,
+    options?: OptionsWithRql
+  ): AsyncGenerator<PagedResult<Document<CustomData>>, Record<string, never>>;
   /**
    * Shortcut method to find a document by id
    *
@@ -919,9 +952,9 @@ export interface DataSchemasService {
    * none | | Every one can use this endpoint
    * `DISABLE_SCHEMAS` | `global` | Includes disabled schemas in the response
    * @param rql Add filters to the requested list.
-   * @returns PagedResult<Schema>
+   * @returns PagedResultWithPager<Schema>
    */
-  find(options?: OptionsWithRql): Promise<PagedResult<Schema>>;
+  find(options?: OptionsWithRql): Promise<PagedResultWithPager<Schema>>;
   /**
    * Request a list of all schemas
    *
@@ -935,6 +968,19 @@ export interface DataSchemasService {
    * @returns Schema[]
    */
   findAll(options?: OptionsWithRql): Promise<Schema[]>;
+  /**
+   * Request a list of all schemas and returns a generator
+   *
+   * Permission | Scope | Effect
+   * - | - | -
+   * none | | Every one can use this endpoint
+   * `DISABLE_SCHEMAS` | `global` | Includes disabled schemas in the response
+   * @param rql Add filters to the requested list.
+   * @returns Schema[]
+   */
+  findAllIterator(
+    options?: OptionsWithRql
+  ): AsyncGenerator<PagedResult<Schema>, Record<string, never>, void>;
   /**
    * Find By Id
    * @param id the Id to search for