+++ tests/unit/data/schemasService.test.ts
@@ -1,6 +1,11 @@
 import nock from 'nock';
 import { AUTH_BASE, DATA_BASE } from '../../../src/constants';
-import { Client, createClient, ParamsOauth2 } from '../../../src/index';
+import {
+  Client,
+  createClient,
+  ParamsOauth2,
+  rqlBuilder,
+} from '../../../src/index';
 import {
   newSchemaInput,
   newSchemaCreated,
@@ -150,5 +155,94 @@ describe('Schemas Service', () => {
     } = await sdk.data.schemas.find();
     const transition = schema.transitionsByName.move;
     expect(transition.id).toBe('5e9fff9d84820a2a9a718e2f');
+  });
+
+  it('should request a list of all schemas via iterator', async () => {
+    nock(`${host}${DATA_BASE}`)
+      .get('/?limit(50)')
+      .reply(200, {
+        page: {
+          total: 55,
+          offset: 0,
+          limit: 50,
+        },
+        data: Array(50).fill(schemaData),
+      })
+      .get('/?limit(50,50)')
+      .reply(200, {
+        page: {
+          total: 55,
+          offset: 50,
+          limit: 50,
+        },
+        data: Array(5).fill(schemaData),
+      });
+    const schemas = sdk.data.schemas.findAllIterator();
+
+    await schemas.next();
+    const thirdPage = await schemas.next();
+    expect(thirdPage.value.data.length).toBe(5);
+  });
+
+  it('should request a list of all schemas with offset and throw error', async () => {
+    expect.assertions(1);
+    try {
+      nock(`${host}${DATA_BASE}`)
+        .get('/?limit(50)')
+        .reply(200, schemasListResponse);
+      await sdk.data.schemas.findAll({
+        rql: rqlBuilder().limit(20, 20).build(),
+      });
+    } catch (error) {
+      expect(error).toBeDefined();
+    }
+  });
+
+  it('should request a list of schemas and use next / previous', async () => {
+    nock(`${host}${DATA_BASE}`)
+      .get('/')
+      .reply(200, {
+        page: {
+          total: 45,
+          offset: 0,
+          limit: 20,
+        },
+        data: Array(20).fill(schemaData),
+      })
+      .get('/?limit(20,20)')
+      .reply(200, {
+        page: {
+          total: 45,
+          offset: 20,
+          limit: 20,
+        },
+        data: Array(20).fill(schemaData),
+      })
+      .get('/?limit(20,40)')
+      .reply(200, {
+        page: {
+          total: 45,
+          offset: 40,
+          limit: 20,
+        },
+        data: Array(5).fill(schemaData),
+      })
+      .get('/?limit(20,20)')
+      .reply(200, {
+        page: {
+          total: 45,
+          offset: 0,
+          limit: 20,
+        },
+        data: Array(20).fill(schemaData),
+      });
+    const schemas = await sdk.data.schemas.find();
+
+    const secondPage = await schemas.next();
+    const thirdPage = await schemas.next();
+    const previousPage = await schemas.previous();
+    expect(secondPage.data.length).toBe(20);
+    expect(thirdPage.data.length).toBe(5);
+    expect(previousPage.data.length).toBe(20);
   });
 });